+++ src/methods/$batch/__tests__/async-batch.auto.spec.ts
@@ -8,27 +8,28 @@
 
 /* eslint-disable no-unused-vars,import/no-duplicates,no-constant-condition */
 
-import { asyncBatch, asyncToArray, range } from '../../..';
+import { asyncBatch, range } from '../../..';
+import { asyncUnwrapDeep as asyncUw } from '../../../__tests__/async-helpers';
 
 describe('asyncBatch', () => {
   it('returns an iterable with batches', async () => {
     const iter = asyncBatch(2, [1, 2, 3, 4, 5, 6, 7, 8, 9]);
-    expect(await asyncToArray(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8], [9]]);
+    expect(await asyncUw(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8], [9]]);
   });
 
   it('returns an iterable with batches when passed an iterable', async () => {
     const iter = asyncBatch(2, range({ start: 1, end: 10 }));
-    expect(await asyncToArray(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8], [9]]);
+    expect(await asyncUw(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8], [9]]);
   });
 
   it('returns an iterable with batches when passed an iterable (2)', async () => {
     const iter = asyncBatch(2, range({ start: 1, end: 9 }));
-    expect(await asyncToArray(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8]]);
+    expect(await asyncUw(iter)).toEqual([[1, 2], [3, 4], [5, 6], [7, 8]]);
   });
 
   it('returns an iterable with batches (curried version)', async () => {
     const iter = asyncBatch(2);
-    expect(await asyncToArray(iter(range({ start: 1, end: 10 })))).toEqual([
+    expect(await asyncUw(iter(range({ start: 1, end: 10 })))).toEqual([
       [1, 2],
       [3, 4],
       [5, 6],
@@ -38,7 +39,7 @@ describe('asyncBatch', () => {
   });
 
   it('returns an empty iterable when passed null', async () => {
-    expect(await asyncToArray(asyncBatch(2, null))).toEqual([]);
+    expect(await asyncUw(asyncBatch(2, null))).toEqual([]);
   });
 
   it('errors when passed size <= 0', async () => {